itfMesManager: FUN_003a2660 long->int on lVar3 (FUN_0035f140 returns s32); diff 175->104

+++ src/itfMesManager.c
@@ -669,7 +669,7 @@ u32 FUN_003a2660(void)
 
   u32 uVar2;
 
-  long lVar3;
+  int lVar3;
 
   u64 uVar4;
 
@@ -685,7 +685,7 @@ u32 FUN_003a2660(void)
 
   else {
 
-    iVar1 = *(int *)(DAT_00959eec_abs + (int)lVar3 * 0xd);
+    iVar1 = *(int *)(DAT_00959eec_abs + lVar3 * 0xd);
 
     if (iVar1 == 0) {
